🚚 [Move]: Board Provider 群の decorator を storybook/ サブディレクトリへ分離
decorator.tsx が Provider 本体 (index.tsx) と同階層にあると Storybook
専用 helper と気付きにくいため、__tests__/ と同様に Provider ごとの
storybook/ サブディレクトリへ移動して本実装から物理的に分離する。
BoardCardProvider / BoardColumnProvider / BoardProviders の 3 つを
同時に揃え、参照する 4 つの Story ファイルの import パスも更新する。
